:white_check_mark: test(Employee): 특정 사원을 id 기준으로 조회

+++ src/main/java/com/api/ecoassignment/domain/employee/dao/EmployeeRepository.java
@@ -3,5 +3,6 @@ package com.api.ecoassignment.domain.employee.dao;
 import com.api.ecoassignment.domain.employee.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee,Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+
 }
